Made loop variable long in case of bugs

+++ src/NumberTheory/PrimeDecomposition.java
@@ -11,9 +11,10 @@ public class PrimeDecomposition {
 			res.add((long) 2);
 			a = a / 2;
 		}
-		for (int i = 3; i <= Math.sqrt(a); i = i + 2) {
+		// Careful! Loop variable is a long
+		for (long i = 3; i <= Math.sqrt(a); i = i + 2) {
 			while (a % i == 0) {
-				res.add((long) i);
+				res.add(i);
 				a = a / i;
 			}
 		}
